requirements.txt
alabaster==0.7.6
Babel==2.1.1
binaryornot==0.4.0
chardet==2.3.0
click==5.1
cookiecutter==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-allauth==0.23.0
django-autoslug==1.8.0
django-braces==1.8.1
django-crispy-forms==1.5.2
django-environ==0.3.0
django-floppyforms==1.5.2
django-model-utils==2.3.1
django-secure==1.0.1
django-social-auth==0.7.28
docutils==0.12
funcsigs==0.4
future==0.15.2
gunicorn==19.3.0
httplib2==0.9.2
Jinja2==2.8
MarkupSafe==0.23
mock==1.3.0
model-mommy==1.2.5.1
mongoengine==0.10.0
oauth2==1.9.0.post1
oauthlib==1.0.3
pbr==1.8.1
Pillow==2.9.0
psycopg2==2.6.1
Pygments==2.0.2
PyJWT==1.4.0
pymongo==3.1
python-openid==2.2.5
python-social-auth==0.2.13
pytz==2015.4
requests==2.8.1
requests-oauthlib==0.5.0
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.10.12
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
static3==0.6.1
unicode-slugify==0.1.3
Unidecode==0.4.18
whichcraft==0.1.1
whitenoise==2.0.3
